feat(database): Implementation of a new database factory
